pyproject.toml
[project]
name = "notebooklm-repo-artefacts"
version = "0.1.0"
description = "Generate NotebookLM artefacts (audio, video, slides, infographic) from any git repository"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "NetDevAutomate"}]
keywords = ["notebooklm", "repository", "documentation", "audio", "video", "slides", "infographic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "typer>=0.12",
    "md2pdf-mermaid>=1.4",
    "notebooklm-py[browser]>=0.3,<0.4",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/NetDevAutomate/notebooklm-repo-artefacts"
Repository = "https://github.com/NetDevAutomate/notebooklm-repo-artefacts"
Issues = "https://github.com/NetDevAutomate/notebooklm-repo-artefacts/issues"

[project.scripts]
repo-artefacts = "repo_artefacts.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/repo_artefacts"]

# --- Dependency Groups (PEP 735) ---

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
lint = [
    "ruff>=0.15.5",
    "pyright>=1.1.408",
]
dev = [
    {include-group = "test"},
    {include-group = "lint"},
]

# --- Tool Configuration ---

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "A", "RUF", "ASYNC"]
ignore = [
    "E501",      # line length handled by formatter
    "B008",      # function call in default arg (typer pattern)
    "ASYNC109",  # timeout param name (our timeout is for polling, not asyncio.timeout)
    "ASYNC240",  # pathlib in async (single mkdir call, not I/O-bound)
    "RUF001",    # ambiguous unicode (Rich markup uses special chars intentionally)
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.pyright]
pythonVersion = "3.11"
pythonPlatform = "All"
include = ["src"]
exclude = ["**/__pycache__", ".venv", "build", "dist"]
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
reportMissingImports = "error"
reportMissingTypeStubs = "warning"
reportUnusedImport = "error"
reportUnusedVariable = "warning"
